skills/nano-banana/requirements.txt
google-genai
Pillow
pyyaml
